backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.5.30
charset-normalizer==2.0.6
crypto==1.4.1
distlib==0.3.3
Django==3.2.9
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.3.2
idna==3.2
Naked==0.1.31
Padding==0.5
platformdirs==2.4.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
tinyec==0.3.1
urllib3==1.26.7
virtualenv==20.10.0
